refactor: update list component and styling
This commit refactors the list component and related styling.

Changes made:

- Add role prop to List component
- Update ListItemProps interface
- Set default list role to 'list'
- Pass role to rendered list element

- Update CSS selectors to use = instead of ~=
- Adjust font family variable syntax
- Refactor button selectors and style variables

The List component was updated to accept a role prop and pass it to the rendered list element. The default role is now 'list'. ListItemProps was updated accordingly.

The CSS was refactored in several ways:

- Attribute selectors now use = instead of ~=
- Font family variable syntax was adjusted
- Button selectors and style variables were cleaned up

This refactors the list component and styling to follow best practices and improve maintainability.